load the intersphinx builder only when needed through an entrypoint
